restructured ccakem files in repo
